Parser friendly, more complete logs
Reaper status logs are printed even if we are killing a worker, and they
show information for each worker. Information is presented in a key=value
fashion which is way easier to parse, and correctly handled by most log
analysis tools.

Since this is breaking former log parsing, I've taken the liberty to
also clarify logs, showing that master is included in the total.

Signed-off-by: Ulysse Buonomo <buonomo.ulysse@gmail.com>